Add ICE requirement test for func parameters
Add a test checking whether invocations specifying the sizes of arrays
in function parameters should be marked as being invoked where an ICE is
required. Intuitively one would think that such invocations should be
required to be ICEs, but compilers don't actually require this since
arrays in function parameters are implicitly converted to pointers
anyway.