Remove `skipImplicitAndParens()`
Remove `skipImplicitAndParens()` since this was causing Maki to report
false negatives for some properties when the statements that aligned
with macro arguments or expansions were wrapped in parenthesized
expressions. Either replace these calls with calls to the
`IgnoreImplicit()` method, or just completely remove them in favor of
more conservative checks.