Make anonymous type checks more strict
Make anonymous type checks more strict such that they return true if a
macro's body, arguments, or subexpressions reference an anonymous type,
even if that type is hidden behind a typedef.

Note: We may be able to make anonymous types defined inside typedefs a
special case, and not consider them anonymous, since I believe it would
not change the behavior of any C program to treat an unnamed type
defined under a typedef as having the same name as the typedef it is
defined under.