Use tokens to get definition end location
Use a macro definition's last token's end location for the macro's end
definition location instead of just using MI.getDefinitionEndLoc().

If the macro has no tokens then we can just call the
getDefinitionEndLoc() method to get the end of the macro's definition.
If the macro has tokens, however, then we can't use this method because
it doesn't return correct results when a macro contains a
backslash-newline immediately followed by a non-whitespace token. For
instance:

 #define PLATFORM \
 "LINUX"

Here, getDefinitionEndLoc() would say that the macro's definition ends
on line 1, whereas it really ends on line 2. Instead, we can use the
end location of the macro's last token to get this information.