Add property DoesBodyEndWithCompoundStmt
Add the syntactic property DoesBodyEndWithCompoundStmt to check if a
macro expands to a statement that ends with a compound statement. Macros
that satisfy this property do not need to be followed by a semicolon at
their callsites, so translation tools like MerC may wish to avoid
translating them to functions since function calls must be followed by
semicolons at their callsites (assuming they are not part of a larger
expression, which macros that end with a compound statement will not
be).