Rework type-checking API
Rework functions for checking type predicates such that the functions
are divided into two categories: those that check whether a predicate
holds for the most desugared version of a given type, and those that
check if a predicate holds for any level of desugaring of a given type.
Currently, the only type in the latter category is
isAnyDeusgaredTypeDefinedAfter().

This change is to make the type predicate-checking functions more robust
(e.g., to correctly and recursively check for predicates in typeof
types), and to make the code easier to maintain.